Moved tests to correct coverage report.
