+++ src/test/java/ood/interview/cards/SingleDeckTest.java
@@ -1,4 +1,4 @@
-package test;
+package ood.interview.cards;
 
 import static org.junit.jupiter.api.Assertions.*;
